requirements.txt
agno
fpdf
numpy
openai
openpyxl
pandas
pillow
plotly
pymongo
pypdf
python-dotenv
requests
spire.xls
kaleido
scipy